dev!: Reorganize core `utils` into type-based subdirectories.
