+++ .changeset/salty-moments-yell.md
@@ -1,0 +1,5 @@
+---
+"@snapwp/core": minor
+---
+
+dev!: Reorganize core `utils` into type-based subdirectories.
