+++ packages/core/src/utils/index.ts
@@ -1,20 +1,28 @@
 export { default as generateGraphqlUrl } from './generate-graphql-url';
-export { default as getStyleObjectFromString } from './get-style-object-from-string';
-export { default as getClassNamesFromString } from './get-class-names-from-string';
-export { default as getSpacingPresetCssVar } from './get-spacing-preset-css-var';
-export { default as getStylesFromAttributes } from './get-styles-from-attributes';
-export { default as cn } from './cn';
-export { default as isValidUrl } from './is-valid-url';
-export { default as getImageSizeFromAttributes } from './get-image-size-from-attributes';
-export { default as getPxForSizeAttribute } from './get-px-for-size-attribute';
 export { default as parseExternalRemotePatterns } from './parse-external-remote-patterns';
-export { default as findElementAndGetClassNames } from './find-element-and-get-class-names';
-export { default as getColorClassName } from './get-color-class-name';
+
+// Style Attribute utilities
+export { default as cn } from './styles/cn';
+export { default as findElementAndGetClassNames } from './styles/find-element-and-get-class-names';
+export { default as getClassNamesFromString } from './styles/get-class-names-from-string';
+export { default as getColorClassName } from './styles/get-color-class-name';
+export { default as getImageSizeFromAttributes } from './styles/get-image-size-from-attributes';
+export { default as getPxForSizeAttribute } from './styles/get-px-for-size-attribute';
+export { default as getSpacingPresetCssVar } from './styles/get-spacing-preset-css-var';
+export { default as getStyleObjectFromString } from './styles/get-style-object-from-string';
+export { default as getStylesFromAttributes } from './styles/get-styles-from-attributes';
+
+// URL utilities
 export {
 	addTrailingSlash,
 	removeTrailingSlash,
 	addLeadingSlash,
 	removeLeadingSlash,
-} from './url-slash-modify';
-export { isWPSiteUrl, isWPHomeUrl, isInternalUrl } from './url-checker';
-export { toFrontendUri } from './to-frontend-uri';
+	toFrontendUri,
+} from './url/transformers';
+export {
+	isWPHomeUrl,
+	isWPSiteUrl,
+	isInternalUrl,
+	isValidUrl,
+} from './url/validators';
