+++ packages/core/src/utils/styles/get-image-size-from-attributes.ts
