+++ packages/core/src/utils/styles/tests/find-element-and-get-class-names.test.ts
@@ -1,4 +1,4 @@
-import findElementAndGetClassNames from '@/utils/find-element-and-get-class-names';
+import findElementAndGetClassNames from '../find-element-and-get-class-names';
 
 describe( 'findElementAndGetClassNames', () => {
 	it( 'returns class names when renderedHtml contains the specified class selector', () => {
